requirements.txt
Flask
numpy
opencv_contrib_python
opencv_python
Pillow
pybase64
Pyrebase4
Requests
